4_with_rdbms_sqlalchemy/requirements.txt
psycopg2-binary==2.9.5
SQLAlchemy==1.4.46
python-dotenv==0.21.0